Update to Line 22 regarding clarification of Mobile Apps distributed via MDM
I have been in conversations with the Mobile PMs regarding the specifics of deploying the Teams Mobile Apps Via MDM. We do not support the distribution of the Mobile Apps via MDM, However once the App has been deployed, on the device we will support it. in other words, we don't support the distribution mechanism. but we will support the app, once it gets there. 

I have been consulting with Avinash Prasada, Pradeep Mettu; Jeff Yin and Mansoor Jafy on this change